Improved code quality for not compiled notices.
1. Moved the HTML to a view file to comply with separation of concerns.
2. Changed the "no go" conditional checks to guard clauses, i.e.
- to better communicate the intent to stop and bail out
- to stop a "no go" from continuing to flow
3. Flipped the order of the conditionals, as get_option() is slower.
4. Renamed both methods to tell us what will happen when called.
