+++ lib/api/compiler/views/not-compiled-notice.php
@@ -1,0 +1,13 @@
+<?php
+/**
+ * View file for the "not compiled" message.
+ *
+ * @package Beans\Framework\API\Compiler
+ *
+ * @since   1.0.0
+ * @since   1.5.0 Moved to view file.
+ */
+
+?>
+<br />
+<span style="color: #d85030;"><?php echo esc_html( $message ); ?></span>
